add validation client and server side using zod

+++ bytegrad/professional-react-and-nextjs/petsoft/src/actions/doCheckoutPet.ts
@@ -2,17 +2,30 @@
 
 import prisma from "@/lib/db"
 import { TErrorMutatePet } from "@/lib/types"
+import { petIdSchema } from "@/lib/validations"
 import { revalidatePath } from "next/cache"
 
-export async function checkoutPet(petId: string) {
+export async function checkoutPet(petId: unknown) {
   let error: TErrorMutatePet, response
+
+  // Run validations
+  const validatedPetId = petIdSchema.safeParse(petId)
+
+  if (!validatedPetId.success) {
+    error = {
+      message: `Failed to validate pet id when checking out pet`,
+    }
+    return { error, response }
+  }
 
   try {
     response = await prisma.pet.delete({
-      where: { id: petId },
+      where: { id: validatedPetId.data },
     })
   } catch (error) {
-    error = { message: `Failed to delete pet from database (Pet ID: ${petId})` }
+    error = {
+      message: `Failed to delete pet from database (Pet ID: ${petId}): \n\n${error}`,
+    }
   }
 
   if (!error) {
